moving EventManager to EventVisualisationView Module moving DataReaderVSD to EventVisualisationDetectors Module
