+++ EventVisualisation/Detectors/CMakeLists.txt
@@ -11,4 +11,5 @@
 o2_add_library(EventVisualisationDetectors
                SOURCES src/DataInterpreterRND.cxx
                        src/DataInterpreterVSD.cxx
+                       src/DataReaderVSD.cxx
                PUBLIC_LINK_LIBRARIES O2::EventVisualisationBase)
