+++ EventVisualisation/Detectors/src/DataReaderVSD.cxx
@@ -13,14 +13,13 @@
 /// \author julian.myrcha@cern.ch
 /// \author p.nowakowski@cern.ch
 
-#include <EventVisualisationBase/DataReaderVSD.h>
+#include <EventVisualisationDetectors/DataReaderVSD.h>
 #include <TSystem.h>
 #include <TEveManager.h>
 #include <TFile.h>
 #include <TPRegexp.h>
 #include <TEveTrackPropagator.h>
 #include <TEveEventManager.h>
-#include <EventVisualisationBase/EventManager.h>
 #include <EventVisualisationBase/EventRegistration.h>
 
 namespace o2
