+++ EventVisualisation/View/include/EventVisualisationView/EventManager.h
